Separate out pub/sub specific code from Redis data handler
